Update temporaldata.py
Hi,
I ran into an issue when using the from_hdf5() method. Some of the sessions I loaded from the perich miller dataset have an empty .attrs dictionary which leads to an error. One possible solution I found is to add this line to check if the dictionary is empty.